Fedora buildslave: Collapse cabal-installs
`cabal install`s should be collapsed into one command so that the
constraint solver can choose a consistent set of packages.

Fixes an instance of this error:

  fatal error: I found a duplicate definition for symbol
     __hscore_S_IXUSR
  whilst processing object file
     /usr/lib64/ghc-7.4.1/directory-1.1.0.2/HSdirectory-1.1.0.2.o

This happens when we link in two different version of the same
package, creating duplicate symbols.

That cannot happen on master, since there we use cabal to determine
one consistent set of dependencies for building and hide all other
packages on the system, but for Ganeti <= 2.13 we don't have that yet.

This change makes the package versions more consitent, fixing the build.

Signed-off-by: Niklas Hambuechen <niklash@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>